Reading Data
We need to tell Firebase to notify us when chat messages arrive. We do
this by adding a callback to the list of chat messages using the on()
method.
This method takes two arguments: the event type and the callback
function. We'll use the 'child_added' event so that we are notified of
the arrival of individual messages.